Add Bullet System
Add Bullet System

+++ src/weapon/Weapon.java
@@ -5,10 +5,10 @@ import utility.ResourceManager.ImageResource;
 public class Weapon {
 
 	private ImageResource imageResourse;
-	private int attackDamage;
-	private int attackSpeed;
+	private double attackDamage;
+	private double attackSpeed;
 
-	public Weapon(ImageResource imageResource, int attackDamage, int attackSpeed) {
+	public Weapon(ImageResource imageResource, double attackDamage, double attackSpeed) {
 		this.imageResourse = imageResource;
 		this.attackDamage = attackDamage;
 		this.attackSpeed = attackSpeed;
